Добавляет разделение скрипта на модули
